+++ lib/scss_lint/linter_registry.rb
@@ -1,4 +1,6 @@
 module SCSSLint
+  class NoSuchLinter < StandardError; end
+
   module LinterRegistry
     @linters = []
 
@@ -7,6 +9,17 @@ module SCSSLint
 
       def included(base)
         @linters << base
+      end
+
+      def extract_linters_from(linter_names)
+        linter_names.map do |linter_name|
+          begin
+            linter_class = linter_name.split('_').map(&:capitalize).join('')
+            Kernel.const_get(linter_class)
+          rescue NameError
+            raise NoSuchLinter.new("Linter #{linter_class} does not exist")
+          end
+        end
       end
     end
   end